HB2-65 move custom Database function to fields.py
